refactor: remove testing operator struct and create a helper where user can put both function to compute logic
