+++ crates/crypto/bls/Cargo.toml
@@ -22,6 +22,5 @@ eigen-utils.workspace = true
 
 [dev-dependencies]
 eigen-testing-utils.workspace = true
-rand = "0.8.4"
 tokio = { workspace = true, features = ["full"] }
 serde_json.workspace = true
