+++ crates/operator/Cargo.toml
@@ -25,6 +25,7 @@ eigen-types.workspace = true
 rust-bls-bn254.workspace = true
 reqwest.workspace = true
 tokio.workspace = true
+rand.workspace = true
 serde_json.workspace = true
 tracing.workspace = true
 rand_core.workspace = true
@@ -34,8 +35,6 @@ eigen-utils.workspace = true
 url.workspace = true
 
 hex = { version = "0.4.3", features = ["alloc"] }
-rand = { version = "0.9.1", features = ["thread_rng"] }
-
 # RPC Server
 tarpc = { version = "0.36", features = ["full"] }
