--- crates/operator/src/testing_operator.rs
@@ -1,197 +1,0 @@
-use crate::{client::ClientAggregator, config, error::OperatorError};
-use alloy::{
-    primitives::keccak256,
-    providers::{Provider, ProviderBuilder, WsConnect},
-    rpc::types::Filter,
-    sol_types::{SolEvent, SolValue},
-};
-use eigen_aggregator::SignedTaskResponse;
-use eigen_client_avsregistry::reader::AvsRegistryChainReader;
-use eigen_crypto_bls::{BlsKeyPair, OperatorId};
-use eigen_logging::logger::SharedLogger;
-use eigen_task_processor::task_response::TaskResponse;
-use futures_util::StreamExt;
-use rand::Rng;
-use serde::{Deserialize, Serialize};
-use tracing::info;
-
-/// USE THIS OPERATOR FOR TESTING PURPOSES ONLY
-/// Testing Operator struct handle the operator logic of processing new tasks
-/// and sending signed task responses to the aggregator. This operator will respond
-/// incorrectly with a chance of `failure_rate`.
-#[derive(Debug)]
-pub struct TestingOperator {
-    operator_id: OperatorId,
-    operator_name: String,
-    client_aggregator: ClientAggregator,
-    ws_rpc_url: String,
-    key_pair: BlsKeyPair,
-    failure_rate: u8,
-}
-
-impl TestingOperator {
-    /// Initialize a new operator.
-    /// This method does not register the operator.
-    ///
-    /// # Arguments
-    ///
-    /// * `key_pair` - The key pair of the operator.
-    /// * `operator_address` - The address of the operator.
-    /// * `operator_name` - The name of the operator.
-    /// * `logger` - The logger.
-    /// * `ws_rpc_url` - The URL of the WebSocket RPC.
-    /// * `http_rpc_url` - The URL of the HTTP RPC.
-    /// * `registry_coordinator_address` - The address of the registry coordinator.
-    /// * `operator_state_retriever_address` - The address of the operator state retriever.
-    /// * `aggregator_ip_port` - The IP and port of the aggregator.
-    /// * `failure_rate` - The chance (0 - 100) that the operator will respond incorrectly.
-    ///
-    /// # Returns
-    ///
-    /// * `Result<Self, OperatorError>` - The operator.
-    pub async fn new(
-        logger: SharedLogger,
-        config: config::TestingOperatorConfig,
-    ) -> Result<Self, OperatorError> {
-        let config::TestingOperatorConfig {
-            bls_key_pair,
-            operator_address,
-            operator_name,
-            ws_rpc_url,
-            http_rpc_url,
-            registry_coordinator_address,
-            operator_state_retriever_address,
-            aggregator_ip_port,
-            failure_rate,
-        } = config;
-        let avs_registry_reader = AvsRegistryChainReader::new(
-            logger,
-            registry_coordinator_address,
-            operator_state_retriever_address,
-            http_rpc_url.to_string(),
-        )
-        .await?;
-
-        let client_aggregator = ClientAggregator::new(aggregator_ip_port).await?;
-
-        let is_registered = avs_registry_reader
-            .is_operator_registered(operator_address)
-            .await
-            .map_err(|_| OperatorError::RegistrationError)?;
-        info!("{} registered: {}", operator_name, is_registered);
-
-        if !is_registered {
-            return Err(OperatorError::RegistrationError);
-        }
-
-        let operator_id = avs_registry_reader
-            .get_operator_id(operator_address)
-            .await
-            .map_err(|_| OperatorError::OperatorIdError)?;
-
-        Ok(Self {
-            operator_id,
-            operator_name: operator_name.to_string(),
-            ws_rpc_url: ws_rpc_url.to_string(),
-            client_aggregator: client_aggregator.clone(),
-            key_pair: bls_key_pair.clone(),
-            failure_rate,
-        })
-    }
-
-    /// Start listening for new task events. Operator subscribe to the event signature of the task processor.
-    /// When a new task is created, the operator will process it and send the signed task response to the aggregator.
-    /// User must to use the structs generated by the bindings because they implement the SolType and SolValue traits.
-    /// This operator will respond with the `failure_logic` with a probability of `failure_rate`.
-    ///
-    /// # Arguments
-    ///
-    /// * `self` - The operator.
-    /// * `compute_logic` - The logic to compute the task response.
-    /// * `failure_logic` - The logic to give a wrong task response.
-    ///
-    /// # Returns
-    ///
-    /// * `Result<(), OperatorError>` - The result of the operation.
-    pub async fn start<Event, FC, FF, Output>(
-        &self,
-        compute_logic: FC,
-        failure_logic: FF,
-    ) -> Result<(), OperatorError>
-    where
-        Event: SolEvent,
-        FC: Fn(Event) -> Result<TaskResponse<Output>, OperatorError>,
-        FF: Fn(Event) -> Result<TaskResponse<Output>, OperatorError>,
-        Output: SolValue + Serialize + for<'de> Deserialize<'de> + Clone,
-    {
-        let ws = WsConnect::new(&self.ws_rpc_url);
-        let provider = ProviderBuilder::new()
-            .on_ws(ws)
-            .await
-            .map_err(|_| OperatorError::TransportError)?;
-
-        let filter = Filter::new().event_signature(Event::SIGNATURE_HASH);
-        let sub = provider
-            .subscribe_logs(&filter)
-            .await
-            .map_err(|_| OperatorError::SubscribeLogsError)?;
-        let mut stream = sub.into_stream();
-
-        while let Some(log) = stream.next().await {
-            let data: Event = log
-                .log_decode()
-                .map_err(|_| OperatorError::SubscribeLogsError)?
-                .inner
-                .data;
-
-            info!("{} picked up a new task", self.operator_name);
-
-            // Generate a boolean with probability of `failure_rate`
-            let mut rng = rand::rng();
-            let should_fail = rng.random_bool(self.failure_rate as f64 / 100.0);
-
-            let task_response: TaskResponse<Output> = if should_fail {
-                info!("{} is failing this task on purpose", self.operator_name);
-                failure_logic(data)?
-            } else {
-                info!("{} is computing the task", self.operator_name);
-                compute_logic(data)?
-            };
-
-            let signed_task_response =
-                Self::sign_task_response(&self.key_pair, &self.operator_id, task_response)?;
-            self.client_aggregator
-                .send_signed_task_response(signed_task_response)
-                .await?;
-        }
-
-        Ok(())
-    }
-
-    /// Sign the task response for the aggregator.
-    ///
-    /// # Arguments
-    ///
-    /// * `key_pair` - The key pair of the operator.
-    /// * `operator_id` - The id of the operator.
-    /// * `task_response` - The task response to sign.
-    ///
-    /// # Returns
-    ///
-    /// * `Result<SignedTaskResponse<Response>, OperatorError>` - The signed task response.
-    fn sign_task_response<Response>(
-        key_pair: &BlsKeyPair,
-        operator_id: &OperatorId,
-        task_response: TaskResponse<Response>,
-    ) -> Result<SignedTaskResponse<TaskResponse<Response>>, OperatorError>
-    where
-        Response: SolValue + Serialize + for<'de> Deserialize<'de> + Clone,
-    {
-        let encoded = task_response.encode();
-        let hash_msg = keccak256(encoded);
-        let signed_msg = key_pair.sign_message(&hash_msg);
-        let signed_task_response = SignedTaskResponse::new(task_response, signed_msg, *operator_id);
-        info!("Operator signed task response");
-        Ok(signed_task_response)
-    }
-}
